Corregido detalle en los graficos de Plan vs Ventas Reales, finalizados los graficos y sevadu, en espera de los requerimientos faltantes....

+++ trunk/bizapps-sevadu/src/servicio/maestros/SPlanVenta.java
@@ -42,7 +42,7 @@ public class SPlanVenta {
 		return planVentaDAO.findByIdMaestroProductoCodigoProductoDusa(clave);
 	}
 
-	public Integer sumarPlanAliado(MaestroAliado maestroAliado,
+	public Double sumarPlanAliado(MaestroAliado maestroAliado,
 			MaestroProducto maestroProducto, int anno, int mes) {
 		return planVentaDAO
 				.sumByIdMaestroAliadoAnIdMaestroProductoAndIdMesAndIdAnno(
